new/requirements.txt

reflex==0.8.17
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.0.0
respx>=0.20.2